skip admin check on Windows to run in elevated mode
on Windows/Hyperv Podman Dekstop needs to run in elevated mode to allow Podman to work (it is not allowed to run Podman as an unprivileged user).
However crc does not work when run as admin. The only way to make it work is to disable the admin check run during the setup/start.
This commit disable this check just before running the setup/start command and restore the original value to avoid maintaining a fixed truevalue. In future Podman would be able to run as unprivileged user and this check could be removed.

Signed-off-by: Luca Stocchi <lstocchi@redhat.com>